Add support for automatically adding known_hosts
Currently, if a server is not in the known_hosts file then
the client will fail with an error like:

paramiko.SSHException: Server '[<host>]:<port>' not found in known_hosts

This change allows users to configure paramiko with the AutoAddPolicy().
It should be noted that using this policy opens users to
man-in-the-middle attacks, which is why it is an optional setting and
defaults to False.